Use neutronclient for floatingip operations
As add_floating_ip and remove_floating_ip were removed from novaclient,
migrating to neutronclient for floatingip operations.

This is the first patch for fixing unittests. As integration tests
didn't fail, floatingip operations are not covered there, so it may be
necessary to add new scenario tests for this in following patches.

Change-Id: Ic41f5898db77a09158e0f8bfa4196bdb5e40b49a
Signed-off-by: Zhao Chao <zhaochao1984@gmail.com>