Register all replicas in replication scenario test
In the replication scenario tests, two replicas will be created by a
instance create with replica_count equals to 2, but only one of them
will be returned in the reponse, so if the testing fails, the debugging
logs of the the other one will not be gathered. This patch will fix.

The 'replica_2_id' property of ReplicationRunner is also removed in this
patch as it's not used at all.

Change-Id: I05fca265e6e5f06b515c009203865f0d48acf4dd
Signed-off-by: Zhao Chao <zhaochao1984@gmail.com>