Add check_process() for mysqldump
Before, mysqldump had redirect parts for stderr but did not have
any check logic for that. So redirect part was deleted to avoid
reporting which is always success.

But, mysqldump does not emit not only errors but also warnings like
"Warning: ..." or "[Warning] ...". This warnings should not be
considered to error, so add check_process for filtering that like
other backup strategies.

Also, change message log level to debug for the right purpose and
change print location not to print None when the output file is empty.

Change-Id: I27b1b0628e79ca1c7b0c1ab5f939b5960ec3177a
Close-Bug: 1756806