Raise timeout for instance resizing checking job
Recently the trove-functional-mysql usually failed by timeout at
the test_instance_returns_to_active_after_resize case, and the cause is
creating and resizing instances could be quit slow(e.g. 1200+ seconds).
This may be caused by the pressure of the CI system and should be
further investigated. The patch is a quick fix for the problem by
raising the timeout to 30 minutes(it's now 900 seconds by default, but
instead it's 32 minutes for instance creating, so raising the value to
30 minutes should be reasonable).

Partial-Bug: #1778837

Change-Id: Ib903a72b098499553978ec881843a7bbee25b45f
Signed-off-by: Zhao Chao <zhaochao1984@gmail.com>