Remove nova conf for guestagent
It seems that the current guestagent only needs to communicate
through amqp, and does not need to interact with nova, so the
related nova configuration can be removed from the guestagent,
which can reduce some invalid configuration.

Similarly, we removed the related configuration of condutor
and guestagent in devstack plugin.

Change-Id: I3f0b3b445c053351c1f3e966d77ea228df0e94a7