Fix the mysql start up issue after restore
When the trove guest image is built without the proper permission
setting for /tmp directory in which ini file is created. (could happen
when using non-official ways of image building), mysqld_safe will fail
to reset the root password after mysql restore due to AppArmour
complaints. As a consequence it fails the whole mysql restore procedure.

Changing init-file ownership fixes this problem and makes the backup and
restore successful.

Co-Authored-By: Jake Yip <jake.yip@unimelb.edu.au>

Closes-Bug: #1791878

Change-Id: Ia164345239197f93fb63697212a9b96c900fbe89