Fix Mariadb replication config
This patch changes name of Mariadb binary and relay log files. From now
they start with 'mariadb' instead of 'mysql'. Thanks to this fact
binary log names on a master and slaves have same name and we avoid
mess in data directory once we promote one of slaves to the master.

Change-Id: I18823b7a78e07b7999c937d068fb9cce9ed00285
Co-Authored-By: Kasper Hasior <k.hasior@samsung.com>