Fix poll_until exception type
This patch adds catching loopingcall.LoopingCallTimeOut and raising
PollTimeOut exceptions.

Without this change we didn't know which exception would be thrown in
case of task timeout in function poll_until. As in the rest of the
trove code we catch only PollTimeOut exception it was possible to miss
loopingcall.LoopingCallTimeOut.

Change-Id: Ic3f8bdf99e82ac5c438354a4f3af7c6856d14f34