Add Cassandra datastore upgrade
This patch adds support for upgrading Cassandra datastore - both single
instance cluster and multiple-node cluster.

It is achievied in a similar way to MySQL/Mariadb. The key difference
is draining nodes before upgrade and upgrading sstables after succesful
volume mounting operation.

Cassandra upgrade steps have been taken from
https://docs.datastax.com/en/upgrade/doc/upgrade/ddac/upgdDDACMinor51x.html

In case of trove it looks as follows:
For each node do:
    + enter restart mode
    + drain and stop node
    + preserve configuration files and home directory
    + unmount volume
    + create instance with new datastore version
    + mount volume
    + restore configuration and home directory
    + start Cassandra
    + upgrade sstables

Because we need to upgrade seeds first there is an additional parameter
in rolling_upgrade_cluster function - a function which is used to so
sort nodes in proper order.

Story: #2005410
Task: #30426
Co-Authored-By: Przemysław Godek <p.godek@partner.samsung.com>
Change-Id: I9f25606d51b35e264a8c36f9524ae57b0a5780d0
Signed-off-by: Kasper Hasior <k.hasior@samsung.com>