Changing file owner when upgrading mariadb
When we upgrade mariadb version we mount volumes with data and
config files.

As we use new images we cannot make assumption that UIDs are the same
as on the old image so we set a new files and directories owner.

Two new functions have been added: _restore_directory() and
_restore_home_directory() in class Manager which are responsible to
copy files from the volume to the instance file system and to change
owner and group of the files.

The new function is located in base Manager class as it very likely
that other datastores may use it as well.

Co-Authored-By: Przemysław Godek <p.godek@partner.samsung.com>
Change-Id: I3c39f51b471081eeabed55070dc91807544a2dda
Story: #2005398
Task: #30391
Signed-off-by: Kasper Hasior <k.hasior@samsung.com>