Remove the definition of trove-grenade
(and of its now unused parent trove-legacy-dsvm-base)

The job is not used in this repository anymore (since
I49bd6c824b837d22d9f3945a3e8c2343c831b352).
It was only used by grenade.git, but that usage has been removed
by I8f976041e98502de4770ece8cb7f04a928e8f3d9.

Moreover, it is a legacy job and it should be ported to a native
Zuul v3 job, and more important it doesn't really test the upgrade,
because there are no grenade plugins for trove.

An upgrade job is useful, but a grenade trove plugin should be written
first and a new native job should be created instead of restoring
the legacy one.

Depends-On: https://review.opendev.org/749212
Change-Id: Ib5ecc496aee28a6a2bbb01138e1f511bfc179302