Add taskmanager_manager option to common
Option `taskmanager_manager` does not have defaults and thus is required
to be set, otherwise trove-taskmanager will fail to start with
AttributeError. In the meanwhile this is not clear what value is expected
as option is not described anywhere.

This change moves option to the general list, which will make it
documented along with setting reasonable default.

Change-Id: I546ffe88b6e135c3f4c6f71b5aef03b92011bed8