instance model: fix extraction of ip addresses from instance record
The instance.addresses value will be a dictionary of
network: [addr_info] key/value pairs. addr_info is not
itself the address but a dict that contains the address
under ['addr'].

The nova api changed quite a while ago so we need to
catch up.

In addition: this function previously returned a list of
dictionaries; users of the function are clearly expecting
a list of simple IPs, so return that instead.

Change-Id: I9f7fed17b93e909f572949037b4e229d527e8fe4