Don't create nova instances with fqdn as their name
If dns integration is enabled, a db instance's 'hostname'
refers to the dns-determined fqdn of a database instance. It will
probably be something like whatever.trove.example.org

Passing that fqdn into nova gets us a VM whose hostname
is whatever.trove.example.org. Presumably that host is itself
in a domain, so we wind up with an fqdn on the VM of something like
whatever.trove.example.org.example.org. Seems wrong.

Furthermore, if this is running in a cloud that already has
automatic DNS integration (a standard neutron/designate thing)
then we wind up trying to create a record in the 'example.org'
domain with the record name of whatever.trove.example.org
which is likely to confuse designate along with human
users.

Story: 2008915
Task: 42506
Change-Id: I6d223f9d1a15a1ed0a0269730a544c321cf6595f