Removes the deprecated argument tenant from TroveContext
The tenant argument of RequestContext in oslo.context had been
deprecated long time ago and it was finally removed in
oslo.context-4.0.0. We should remove the tenant argument of
TroveContext that derives from RequestContext and we should also
update the requirements.txt in the master branch.

The original problem of this patch is the Configurations API
errors in following cases:
* when getting a configuration.
* when adding a new configuration parameter.

Story: 2009933
Task: 44816
Change-Id: Idbd0507e6c87d6926f93806d0d1eef13f557d805