Don't check task_status when creating a cluster
This PR allows load_simple_instance_addresse to get the ipaddr of an
instance from neutron when creating a instance belongs to a cluster by
skipping checking the task_status.

load_simple_instance_addresses, which is a helper function to load ip
addr information to Instance instances, currently checks task_status
before getting the information from neutron. However, this logic is
unnecessary and causes a logic error when instances belong to a
cluster because the task_status of creating a cluster is determined
after initializing a cluster, which happens in the next phase of
load_simple_instance_addresses.

The original reason of this code change is that the task status of a
mariadb cluster hasn't change from BUILDING to NONE even after
successfully creating instances belong to the cluster.

Task: 45110
Story: 2009982
Change-Id: Id3234081fd3114e6536358cbdbf94b8280c8bc41