Adds new configurations for injected file owner
This PR adds configuration options for file owner information of
injected files in GuestVM to trove.conf. File owner information
should not be hardcoded as "trove" because it depends on user
environments.

Original problem is guest instance creation failure due to the lack of
guestagent configuration files because of the recent code chage[1].
That patch was introduced since Yoga.

[1]:https://review.opendev.org/c/openstack/trove/+/823632

Task: 45239
Story: 2010021
Change-Id: Iaccfe84c912025ec6d96e66ffa6ad7a4b9537b6e