Add UID/GID configurations for each datastore
When trove prepares a container, it grants permissions for folders with
the default user/group "database" (ID 1001). However, some Docker images
require a different user for provisioning the database. Therefore, we
need to make it flexible so that the database container can run with
the required user and group, regardless of the default settings.

Story: #2010827
Task: #48386
Change-Id: Ie0268951db7a9e711089714cecf7fbf311eecec8