Fix potential network conflict
when both the management and business network have a gateway.
The guest vm may generate an unexpected default gateway
which is the businees gateway as opposed to the management one.
In this case, trove guest VM may failed to connect with the
openstack control plane such as swift service while doing the
backup restore action.
In addition, if the mgmt port and the business port have the same
IP address. This also may cause network conflict

This change disable the user defined port once the guest-agent
starts when the network isolation is enabled.

Change-Id: I7a96952f34ce5f4aead837b94daedd83c0a871d8