Use policy to determine admin context
Currently the logic to determine the admin context is controlled by
the [DEFAULT] admin_roles option but this requires that users set both
policy and the option properly.

This replaces the option by a new policy rule so that users can define
all policy related settings in the policy.yaml file. This also allows
users to create their is_admin rule with not only roles but also with
other elements.

Change-Id: Ib9d003065f943a5382e1079397bfde4292f5dabb