[stable-only] Cap setuptools <71.0.0
py38 jobs (on ubuntu-focal) started to fail due to recent virtualenv
release (20.26.4) on Yoga (which bundles setuptools), because we have
'packaging==21.3' in this branch that is not compatible with newer
setuptools [1].

setuptools is bundled in virtualenv, so it has to be capped via the
virtualenv package. Note, that we also have to cap it explicitly,
otherwise tox somehow pulls in latest setuptools.

[1] https://github.com/pypa/setuptools/issues/4483

Change-Id: Id9e02f8b26bad5e2ef74908012de0b81898c04d0