Add build of modern Ubuntu distributions
With world moving forward, we need to assure that more modern
tools and images are going to be built for DB Backups.

For that we are starting with allowing to provide a different OS
for the Dockerfile as well as add OS version to the promote
jobs.

This should be the first step before moving forward with adding more
jobs/images  on modern distributions.
Change preserves current behaviour, as new ubuntu images need to be
published before proceeding further.

Change-Id: I7b4fc6ed8771f119eaac19d8643e78d2a36bb02f