Fail the backup properly if no backup image is supplied
At the moment if no backup image is supplied in configuration,
the creation will fail with AttributeError on _image_has_tag
verification as it will attempt to split NoneType.

With that, such backup will stuck it the New state, as
get_backup_image is called from outside of try/except. This also
blocks any further attempts to create backup as well as does not
provide any result to the user, and requires a manual intervention
into database to reset the backup state to Error.

This patch fix error handling as well as add a human-readable error
when configuration does not contain backup images.

Change-Id: Ia9a0d032033a6f99c2faf42a7f8c79214b247904