Allow to have less specific versioned backup images
In most usecases operator does not have a need to have an image
per each specific database version. Backup images are suitable for
the whole series of versions in most cases.

Mor instance, backup image tagged as 10.4 should be working
nicely for 10.4.1, 10.4.2, 10.4.3, etc. This patch adds iteration
and search in registry for images before attempting to create
a container from it.

In case backup image is not found, we will remove a minor or
bugfix version and try again.

If no image will be found at the end - TroveError will be raised.

Change-Id: Ib0394004e673fa8458b87b21b394c2af24c5e9ce
Signed-off-by: Dmitriy Rabotyagov <noonedeadpunk@gmail.com>