Fix: Allow float values in KeyValueCodec
Fix "AttributeError: 'float' object has no attribute
'lower'" in KeyValueCodec when attempting to set a float
value for a configuration parameter with "float" type.

The `is_valid_value` method was incorrectly attempting to
call `.lower()` on a float, causing a crash. This patch
adds an explicit check for floats to prevent this error.

Change-Id: I4abee725caa762d031172d484ea93fb5fd34d6c4
Signed-off-by: Erkin Mussurmankulov <mangust404@gmail.com>